feat: camelCase on entities and fix AlertController get all method

+++ src/main/java/com/empresa/projeto/gestao_ubs/Dto/Report/DepartmentExpense/DepartmentExpenseDto.java
@@ -6,5 +6,5 @@ public record DepartmentExpenseDto (
     Long department_id,
     String department_name,
     BigDecimal monthly_budget,
-    Double total
+    BigDecimal total
 ){}
