+++ src/main/java/com/empresa/projeto/gestao_ubs/Dto/Report/EmployeeExpense/EmployeeExpenseReportDto.java
@@ -1,10 +1,11 @@
 package com.empresa.projeto.gestao_ubs.Dto.Report.EmployeeExpense;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record EmployeeExpenseReportDto(
         Long employee_id,
         String employee_name,
-        Double total_amount,
+        BigDecimal total_amount,
         List<EmployeeExpenseDto> expenses
 ) {}
